window_server: change wid_t from uint64_t to int

+++ userland/window_server/include/window_server/message.h
@@ -9,7 +9,7 @@
 #include <string.h>
 #include <kernel/hal/input.h>
 
-typedef uint64_t wid_t;
+typedef int wid_t;
 
 namespace WindowServer {
 enum WindowType {
